These changes fix the Xorg bug where the file descriptor for the touchscreen input drivers gets closed but the input driver doesn't get notified when the focus comes back to the Xorg terminal.
I asked Peter Hutterer to file a bug so this will be fixed, but
in the meantime this seems to work. I have basically neutered the
input driver by not closing the file descriptor on both DEVICE_OFF
and DEVICE_CLOSE. I'm not sure what ramifications this could have
in different scenarios, but hopefully this will allow Yum to work
with the current driver on 18.04.2 of their backoffice image.